Move css loader to webpack common config
