+++ config/webpack.dev.js
@@ -5,52 +5,6 @@ const ignoredFiles = require('react-dev-utils/ignoredFiles');
 const common = require('./webpack.common');
 const paths = require('./paths');
 
-process.env.NODE_ENV = 'development';
-
-const regexpStyle = /\.(css|less|styl|scss|sass|sss)$/;
-
-const cssModule = ({ modules = false }) => [
-  {
-    loader: require.resolve('css-loader'),
-    options: {
-      importLoaders: 2,
-      camelCase: true,
-      modules,
-      localIdentName: '[name]-[local]-[hash:base64:5]'
-    }
-  },
-  {
-    loader: require.resolve('postcss-loader'),
-    options: {
-      ident: 'postcss',
-      config: {
-        ctx: {
-          autoprefixer: {
-            browsers: [
-              '>1%',
-              'last 4 versions',
-              'Firefox ESR',
-              'not ie < 10' // React doesn't support IE9 anyway
-            ],
-            flexbox: 'no-2009'
-          }
-        }
-      },
-      plugins: () => [
-        require('postcss-flexbugs-fixes'),
-        require('precss'),
-        require('autoprefixer')
-      ]
-    }
-  },
-  {
-    loader: require.resolve('sass-loader'),
-    options: {
-      sourceMap: true
-    }
-  }
-];
-
 module.exports = merge(common, {
   mode: 'development',
   bail: true,
@@ -69,20 +23,6 @@ module.exports = merge(common, {
     historyApiFallback: true,
     disableHostCheck: true,
     clientLogLevel: 'warning'
-  },
-  module: {
-    rules: [
-      {
-        test: regexpStyle,
-        exclude: paths.appGlobalStyles,
-        use: [require.resolve('style-loader'), ...cssModule({ modules: true })]
-      },
-      {
-        test: regexpStyle,
-        include: paths.appGlobalStyles,
-        use: [require.resolve('style-loader'), ...cssModule({ modules: false })]
-      }
-    ]
   },
   plugins: [
     new HtmlWebpackPlugin({
