+++ package.json
@@ -8,8 +8,8 @@
     "lint-css": "prettier-stylelint",
     "stylelint-check": "stylelint-config-prettier-check",
     "test": "jest",
-    "start": "webpack-dev-server --progress --config config/webpack.dev.js",
-    "build": "rimraf ./build && webpack --progress --config config/webpack.prod.js"
+    "start": "NODE_ENV=development webpack-dev-server --progress --config config/webpack.dev.js",
+    "build": "rimraf ./build && NODE_ENV=production webpack --progress --config config/webpack.prod.js"
   },
   "repository": {
     "type": "git",
@@ -43,15 +43,20 @@
     "axios": "^0.18.0",
     "babel-polyfill": "^6.26.0",
     "bootstrap": "^4.1.1",
+    "classnames": "^2.2.6",
     "core-js": "^2.5.7",
     "mobx": "^4.2.0",
     "mobx-react": "^5.2.3",
     "react": "^16.4.1",
-    "react-bootstrap": "^0.32.1",
     "react-dom": "^16.4.1",
+    "react-ga": "^2.5.3",
+    "react-hot-loader": "^4.3.3",
+    "react-imported-component": "^4.6.2",
     "react-prop-types": "^0.4.0",
     "react-router": "^4.3.1",
-    "react-router-dom": "^4.3.1"
+    "react-router-dom": "^4.3.1",
+    "reactstrap": "^6.3.0",
+    "uniqid": "^5.0.3"
   },
   "devDependencies": {
     "autoprefixer": "^8.6.5",
@@ -63,13 +68,12 @@
     "babel-plugin-transform-function-bind": "^6.22.0",
     "babel-plugin-transform-react-inline-elements": "^6.22.0",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.13",
-    "babel-polyfill": "^6.26.0",
+    "babel-preset-env": "^1.7.0",
     "babel-preset-react-app": "^3.1.2",
     "babel-preset-stage-2": "^6.24.1",
     "babel-runtime": "^6.26.0",
     "case-sensitive-paths-webpack-plugin": "^2.1.2",
     "circular-dependency-plugin": "^5.0.2",
-    "classnames": "^2.2.6",
     "copy-webpack-plugin": "^4.5.2",
     "css-loader": "^1.0.0",
     "directory-named-webpack-plugin": "^4.0.0",
@@ -79,13 +83,13 @@
     "eslint-loader": "^2.0.0",
     "eslint-plugin-prettier": "^2.6.2",
     "eslint-plugin-react-app": "^3.0.0",
-    "extract-text-webpack-plugin": "^4.0.0-beta.0",
     "file-loader": "^1.1.11",
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
     "identity-obj-proxy": "^3.0.0",
     "image-webpack-loader": "^4.3.1",
     "jest": "^23.3.0",
+    "mini-css-extract-plugin": "^0.4.1",
     "node-sass": "^4.9.2",
     "optimize-css-assets-webpack-plugin": "^4.0.3",
     "postcss": "^6.0.23",
@@ -96,10 +100,6 @@
     "prettier-stylelint": "^0.4.2",
     "raf": "^3.4.0",
     "react-dev-utils": "^5.0.1",
-    "react-ga": "^2.5.3",
-    "react-hot-loader": "^4.3.3",
-    "react-imported-component": "^4.6.2",
-    "reactstrap": "^6.3.0",
     "rimraf": "^2.6.2",
     "sass-loader": "^7.0.3",
     "style-loader": "^0.21.0",
@@ -112,9 +112,9 @@
     "stylelint-webpack-plugin": "^0.10.5",
     "svg-url-loader": "^2.3.2",
     "uglifyjs-webpack-plugin": "^1.2.7",
-    "uniqid": "^5.0.3",
     "url-loader": "^1.0.1",
     "webpack": "^4.15.1",
+    "webpack-bundle-analyzer": "^2.13.1",
     "webpack-cli": "^3.0.8",
     "webpack-dev-server": "^3.1.4",
     "webpack-merge": "^4.1.3",
